# sdk/go/api/openapi.yaml
openapi: 3.1.0
info:
  description: "C2PA-compliant content signing and verification infrastructure for\
    \ publishers, legal/finance firms, AI labs, and enterprises"
  title: Encypher Enterprise API
  version: 1.0.0-preview
servers:
- description: Production
  url: https://api.encypherai.com
- description: Local development
  url: http://localhost:8007
paths:
...
  /api/v1/stream/sign:
    post:
      description: Stream signing progress via SSE.
      operationId: stream_signing_api_v1_stream_sign_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamSignRequest"
        required: true
      responses:
        "200":
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Stream Signing
      tags:
      - Streaming
  /api/v1/stream/session/create:
    post:
      description: "Create a new streaming session.\n\nArgs:\n    session_type: Type\
        \ of session (websocket, sse, kafka)\n    metadata: Optional session metadata\n\
        \    signing_options: Optional signing configuration\n    organization: Authenticated\
        \ organization\n    db: Database session\n    \nReturns:\n    Session creation\
        \ result with session_id"
      operationId: create_streaming_session_api_v1_stream_session_create_post
      parameters:
      - explode: true
        in: query
        name: session_type
        required: false
        schema:
          default: websocket
          title: Session Type
          type: string
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Body_create_streaming_session_api_v1_stream_session_create_post"
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Streaming Session
      tags:
      - Streaming
  /api/v1/stream/session/{session_id}/close:
    post:
      description: "Close a streaming session.\n\nArgs:\n    session_id: Session identifier\n\
        \    organization: Authenticated organization\n    db: Database session\n\
        \    \nReturns:\n    Session closure result with stats"
      operationId: close_streaming_session_api_v1_stream_session__session_id__close_post
      parameters:
      - explode: false
        in: path
        name: session_id
        required: true
        schema:
          title: Session Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Close Streaming Session
      tags:
      - Streaming
  /api/v1/stream/events:
    get:
      description: "Server-Sent Events (SSE) endpoint for unidirectional streaming.\n\
        \nArgs:\n    session_id: Session identifier\n    api_key: API key for authentication\n\
        \    \nReturns:\n    StreamingResponse with SSE events"
      operationId: sse_events_endpoint_api_v1_stream_events_get
      parameters:
      - explode: true
        in: query
        name: session_id
        required: true
        schema:
          title: Session Id
          type: string
        style: form
      - explode: true
        in: query
        name: api_key
        required: false
        schema:
          nullable: true
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Sse Events Endpoint
      tags:
      - Streaming
  /api/v1/stream/stats:
    get:
      description: "Get streaming statistics for organization.\n\nArgs:\n    organization:\
        \ Authenticated organization\n    \nReturns:\n    Streaming statistics"
      operationId: get_streaming_stats_api_v1_stream_stats_get
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Streaming Stats
      tags:
      - Streaming
...
  /api/v1/stream/runs/{run_id}:
    get:
      description: Return persisted streaming run state.
      operationId: get_stream_run_api_v1_stream_runs__run_id__get
      parameters:
      - explode: false
        in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Stream Run
      tags:
      - Streaming
...
components:
  schemas:
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    Body_create_streaming_session_api_v1_stream_session_create_post:
      properties:
        metadata:
          additionalProperties: {}
          nullable: true
        signing_options:
          additionalProperties: {}
          nullable: true
      title: Body_create_streaming_session_api_v1_stream_session_create_post
...
    StreamSignRequest:
      description: Request payload for streaming signing run.
      example:
        document_title: document_title
        run_id: run_id
        text: text
        document_id: document_id
        document_type: article
      properties:
        text:
          description: Content to sign while streaming progress
          title: Text
          type: string
        document_id:
          nullable: true
          type: string
        document_title:
          nullable: true
          type: string
        document_type:
          default: article
          description: Document type metadata
          title: Document Type
          type: string
        run_id:
          nullable: true
          type: string
      required:
      - text
      title: StreamSignRequest
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http